url路由重构

+++ distest-web/src/main/java/com/testwa/distest/server/web/project/controller/ProjectController.java
@@ -5,10 +5,10 @@ import com.testwa.core.base.constant.WebConstants;
 import com.testwa.core.base.controller.BaseController;
 import com.testwa.core.base.form.IDListForm;
 import com.testwa.core.base.form.IDForm;
-import com.testwa.core.base.vo.Result;
-import com.testwa.distest.common.util.WebUtil;
+import com.testwa.distest.common.enums.DB;
 import com.testwa.core.base.vo.PageResult;
-import com.testwa.distest.exception.BusinessException;
+import com.testwa.distest.exception.AuthorizedException;
+import com.testwa.distest.server.entity.ProjectMember;
 import com.testwa.distest.server.entity.User;
 import com.testwa.distest.server.entity.Project;
 import com.testwa.distest.server.service.project.form.ProjectListForm;
@@ -59,6 +59,8 @@ public class ProjectController extends BaseController {
     private UserService userService;
     @Autowired
     private ViewMgr viewMgr;
+    @Autowired
+    private User currentUser;
 
     @ApiOperation(value="创建项目")
     @ResponseBody
@@ -68,8 +70,6 @@ public class ProjectController extends BaseController {
         if(form.getMembers() != null && !form.getMembers().isEmpty()){
             userValidator.validateUsernamesExist(form.getMembers());
         }
-
-        User currentUser = userService.findByUsername(WebUtil.getCurrentUsername());
         Project project = projectService.save(form, currentUser);
 
         ProjectVO vo = buildVO(project, ProjectVO.class);
@@ -83,6 +83,9 @@ public class ProjectController extends BaseController {
     @PostMapping(value = "/update")
     public ProjectVO update(@RequestBody @Valid ProjectUpdateForm form) {
         projectValidator.validateProjectExist(form.getProjectId());
+
+        checkProjectAdmin(form.getProjectId());
+
         if(form.getMembers() != null){
             userValidator.validateUsernamesExist(form.getMembers());
         }
@@ -99,16 +102,18 @@ public class ProjectController extends BaseController {
 
     @ApiOperation(value="删除多个项目", notes="")
     @ResponseBody
-    @PostMapping(value = "/delete/all")
+    @PostMapping(value = "/deleteAll")
     public void deleteAll(@RequestBody @Valid IDListForm form) {
+        form.getEntityIds().forEach(this::checkProjectOwner);
         projectService.delete(form.getEntityIds());
     }
 
 
     @ApiOperation(value="删除一个项目", notes="")
     @ResponseBody
-    @PostMapping(value = "/delete/one")
+    @PostMapping(value = "/deleteOne")
     public void deleteOne(@RequestBody @Valid IDForm form) {
+        checkProjectOwner(form.getEntityId());
         projectService.delete(form.getEntityId());
     }
 
@@ -116,7 +121,7 @@ public class ProjectController extends BaseController {
     @ResponseBody
     @GetMapping(value = "/page")
     public PageResult page(@Valid ProjectListForm form){
-        PageResult<Project> projectPR = projectService.findAllByUserPage(form, WebUtil.getCurrentUsername());
+        PageResult<Project> projectPR = projectService.findAllByUserPage(form, currentUser);
         List<Project> projects = projectPR.getPages();
         List<ProjectVO> vos = new ArrayList<>();
         for (Project p : projects) {
@@ -139,9 +144,9 @@ public class ProjectController extends BaseController {
 
     @ApiOperation(value="项目列表")
     @ResponseBody
-    @GetMapping(value = "/select/list")
-    public List selectList() {
-        List<Project> projectsOfUser = projectService.findAllByUserList(WebUtil.getCurrentUsername());
+    @GetMapping(value = "/list")
+    public List list() {
+        List<Project> projectsOfUser = projectService.findAllByUserList(currentUser.getId());
         List<ProjectVO> vos = new ArrayList<>();
         for (Project p : projectsOfUser) {
             User createUser = userService.findOne(p.getCreateBy());
@@ -157,9 +162,11 @@ public class ProjectController extends BaseController {
 
     @ApiOperation(value="获得一个项目的详情（包含成员），同时表明该用户进入了该项目")
     @ResponseBody
-    @GetMapping(value = "/detail/{projectId}")
+    @GetMapping(value = "/{projectId}/detail")
     public ProjectDetailVO detail(@PathVariable Long projectId) throws Exception {
         Project project = projectValidator.validateProjectExist(projectId);
+        projectValidator.validateUserIsProjectMember(projectId, currentUser.getId());
+
         List<User> members = projectMemberService.findAllMembers(projectId);
         viewMgr.setRecentViewProject(projectId);
         User createUser = userService.findOne(project.getCreateBy());
@@ -183,14 +190,32 @@ public class ProjectController extends BaseController {
     @ResponseBody
     @GetMapping(value = "/my/views")
     public List myViews() throws Exception {
-        List<Long> projectIds = viewMgr.getRecentViewProject(WebUtil.getCurrentUsername());
+
+        List<Long> projectIds = viewMgr.getRecentViewProject(currentUser.getUsername());
         List<ProjectVO> vos = new ArrayList<>();
-        if(projectIds != null && projectIds.size() > 0){
+        if(projectIds != null && !projectIds.isEmpty()){
             List<Project> projects = projectService.findByProjectOrder(projectIds);
             vos = buildVOs(projects, ProjectVO.class);
         }
         return vos;
 
+    }
+
+    private void checkProjectOwner(Long entityId) {
+        checkProjectRole(entityId, DB.ProjectRole.OWNER);
+    }
+
+    private void checkProjectAdmin(Long entityId) {
+        checkProjectRole(entityId, DB.ProjectRole.ADMIN);
+    }
+
+    private void checkProjectRole(Long entityId, DB.ProjectRole projectRole) {
+        ProjectMember projectMember = projectMemberService.findByProjectIdAndMemberId(entityId, currentUser.getId());
+        if(projectMember != null) {
+            if (!projectRole.equals(projectMember.getProjectRole())) {
+                throw new AuthorizedException(ResultCode.ILLEGAL_OP, "您无法更改项目");
+            }
+        }
     }
